backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
python-multipart==0.0.9
pdfplumber==0.11.2
python-docx==1.1.2
openai>=1.35.0
python-dotenv==1.0.1
aiofiles==23.2.1
Pillow==10.3.0
pytesseract==0.3.10
